separate bdicwriter files
